refactor utils.py
